Create separate list for commercial libraries.
Commercial libraries are libraries available under a proprietary
license. This includes dual-licensed libraries, which are also
available under an open source license.

Technically open-source libraries can also be commercially used,
but in the Qt world it's common use to refer to proprietary
licenses as the commercial license. So we'll go with what people
will most likely understand here.