Generate manifests from KDE Frameworks git checkout
There is a new command to parse a git checkout and create generic
manifests from the data which is there. It currently parses
README.md and AUTHORS to collect the meta data.

As part of this change there also was added a helper for preparing
given file systems for testing. The GivenFilesystem class makes
it easy to generate hierarchies of directories and files filled
with test data in a way which is suitable for tests. It doesn't
interfere with any other files and it properly cleans up with only
one call of a method.