Use KDE git URL
Inqlude now pulls from the canonical data repository for Inqlude
on the KDE git server. The GitHub repository will continue to
exist to make it easier for non-KDE people to contribute.

This commit also slightly cleans up the pull logic to not delete
existing directories anymore, if they are not a git repository.