Define Manifest attributes programmatically
Add JsonObject. This class dynamically creates a hierarchy of sub
classes from a DSL definition, which reads and writes corresponding
JSON data. With the "attribute" helper the attributes of the class
are defined.

Move Manifest to JsonObject. This simplifies the class significantly
as the attribute only has to defined once with the "attribute"
helper and no other code is necessary to read or write JSON.

Simplify code dealing with Manifests. Some of the code can be
simplified now as the attributes are now accessible as meta data.