Generate links from name
The standard links for frameworks such as home page, mailing list,
git repository always follow the same scheme. So instead of parsing
them from the README.md generate them from the name of the framework.

This is more flexible, allows to change all links at once, and is
compatible with the frameworks where the links have been removed from
the READMEs.