Clarify points of contact
From the experience of the last couple of years it shows that the
mailing list is not used at all by people when contacting us with
questions about Inqlude. All discussions go through issues or
patches, most of them coming via GitHub. To reflect this I'm
removing the mailing list from the web site, so we can close it
down later.